Only warn for failure to resolve client info if player is actually in the game
If the player is kicked before their client can load properly, e.g. if they are banned from the game, Gamebeast will still output a warning because their client info could not resolve. This commit adds a check to make sure the player is actually still in the game before warning to prevent spam in the output.